WIP: Extract app options menu for reuse
